refactor: change recipes folder structure
