+++ frontend/src/components/Recipes/Sidebar/SearchSections/SelectCategory.tsx
