+++ frontend/src/components/Recipes/Sidebar/SearchSections/SelectCuisine.tsx
