+++ frontend/src/components/Recipes/Sidebar/SearchSections/SelectPreference.tsx
@@ -17,4 +17,4 @@ export default function SelectPreference() {
             </select>
         </div>
     )
-} 
+}
