+++ frontend/src/components/Recipes/Sidebar/SearchSections/index.ts
@@ -1,0 +1,6 @@
+import SearchName from './SearchName'
+import SelectCategory from './SelectCategory'
+import SelectCuisine from './SelectCuisine'
+import SelectPreference from './SelectPreference'
+
+export { SearchName, SelectCategory, SelectCuisine, SelectPreference }
