+++ frontend/src/components/Recipes/Sidebar/index.ts
@@ -1,0 +1,13 @@
+import RecipeFilters from './RecipeFilters'
+import IngredientsList from './IngredientsList'
+import UserReviews from './UserReviews'
+import RecipeDifficulty from './RecipeDifficulty'
+import SearchSections from './SearchSections/SearchSections'
+
+export {
+    SearchSections,
+    RecipeFilters,
+    IngredientsList,
+    UserReviews,
+    RecipeDifficulty,
+}
